fix: address google oauth review feedback
