+++ crates/ironclaw_auth/tests/auth_product_contract/oauth_flow_contract.rs
@@ -7,7 +7,6 @@ async fn oauth_callback_exchanges_provider_code_then_completes_once() {
     let flow = oauth_flow(&services, owner.clone()).await;
 
     let request = OAuthProviderCallbackRequest {
-        scope: owner.clone(),
         provider: provider(),
         account_label: label("work github"),
         authorization_code: OAuthAuthorizationCode::new(secret("raw-auth-code"))
@@ -23,7 +22,13 @@ async fn oauth_callback_exchanges_provider_code_then_completes_once() {
     assert!(!debug.contains("raw-pkce-verifier"));
 
     let exchange = services
-        .exchange_callback(request)
+        .exchange_callback(
+            OAuthProviderExchangeContext {
+                scope: owner.clone(),
+                flow_id: flow.id,
+            },
+            request,
+        )
         .await
         .expect("provider exchange");
     let completed = services
