+++ crates/ironclaw_first_party_extensions/src/gsuite/mod.rs
@@ -15,7 +15,7 @@ pub use handlers::{
     GMAIL_GET_MESSAGE_CAPABILITY_ID, GMAIL_LIST_MESSAGES_CAPABILITY_ID,
     GMAIL_REPLY_TO_MESSAGE_CAPABILITY_ID, GMAIL_SEND_MESSAGE_CAPABILITY_ID,
     GMAIL_TRASH_MESSAGE_CAPABILITY_ID, GsuiteCredentialDispatchReason, GsuiteDispatchError,
-    GsuiteDispatchReason, GsuiteDispatchRequest, GsuiteDispatchResult, GsuiteExecutor,
+    GsuiteDispatchRequest, GsuiteDispatchResult, GsuiteExecutor,
 };
 pub use manifest::{
     CALENDAR_EXTENSION_ID, GMAIL_EXTENSION_ID, GSUITE_OUTPUT_BYTES_LIMIT,
