+++ crates/ironclaw_reborn_composition/src/lib.rs
@@ -30,6 +30,7 @@ mod extension_lifecycle;
 mod extension_lifecycle_capabilities;
 mod extension_lifecycle_command;
 mod factory;
+mod google_oauth;
 mod gsuite;
 mod input;
 mod lifecycle;
