+++ crates/ironclaw_reborn_composition/src/product_auth_serve.rs
@@ -543,7 +543,7 @@ async fn oauth_callback_handler(
 fn callback_outcome_from_query(
     state: &ProductAuthRouteState,
     flow_id: AuthFlowId,
-    scope: &AuthProductScope,
+    _scope: &AuthProductScope,
     query: &OAuthCallbackQuery,
 ) -> Result<RebornOAuthCallbackOutcome, ProductAuthRouteFailure> {
     if query
@@ -565,8 +565,7 @@ fn callback_outcome_from_query(
     let pkce_verifier_hash = pkce_verifier_hash(pkce_verifier.expose_secret())?;
 
     Ok(RebornOAuthCallbackOutcome::Authorized {
-        provider_request: Box::new(OAuthProviderCallbackRequest {
-            scope: scope.clone(),
+        provider_request: OAuthProviderCallbackRequest {
             provider: AuthProviderId::new(provider.to_string())
                 .map_err(|_| ProductAuthRouteFailure::malformed_callback())?,
             account_label: CredentialAccountLabel::new(account_label.to_string())
@@ -578,7 +577,7 @@ fn callback_outcome_from_query(
                 .map_err(ProductAuthRouteFailure::from)?,
             pkce_verifier_hash,
             scopes: parse_provider_scopes(query.scopes.as_deref())?,
-        }),
+        },
     })
 }
